requirements.txt
Django==2.2.3
django-ckeditor==5.7.1
django-crispy-forms==1.7.2
django-js-asset==1.2.2
gunicorn==19.9.0
Pillow==6.1.0
pytz==2019.1
sqlparse==0.3.0
whitenoise==4.1.3
